backend/requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
PyYAML==6.0.2
pytest==8.3.3
